feat: impl https proxy for elastic beanstalk deploy
