+++ .env.example
@@ -1,4 +1,5 @@
 NODE_ENV=development
+BEANSTALK_URL=<your-beanstalk-url>
 FRONTEND_ORIGIN=<your-frontend-origin>
 LOCALDEV_ORIGIN=<your-frontend-development-origin>
 DB_NAME=<your-db-name>
@@ -6,6 +7,6 @@ DB_HOST=<your-db-host>
 DB_PORT=<your-db-port>
 DB_USERNAME=<your-db-username>
 DB_PASSWORD=<your-db-password>
-DB_RESOURCE_ID=<your-db-resource-id>
-DB_SECURITY_GROUP_ID=<your-db-security-group-id>
+# DB_RESOURCE_ID=<your-db-resource-id>
+# DB_SECURITY_GROUP_ID=<your-db-security-group-id>
 DB_CERTIFICATE=<optional-certificates>
